+++ JavaProject/Circle.java
@@ -2,7 +2,7 @@
 // ΑΜ: [Ο Αριθμός Μητρώου Σας]
 
 public class Circle extends Shape { //κληρονομει από την κλάση Shape
-    public Circle(int accessMethodForSquare) {
+    public Circle(double accessMethodForSquare) {
         super("Circle", accessMethodForSquare); //κλήση του constructor της υπερκλάσης με τον τύπο "Circle" και το εμβαδόν του περιβάλλοντος τετραγώνου
     }
