+++ JavaProject/Triangle.java
@@ -2,7 +2,7 @@
 // ΑΜ: [Ο Αριθμός Μητρώου Σας]
 
 public class Triangle extends Shape { //κληρονομει από την κλάση Shape
-    public Triangle(int accessMethodForSquare) {
+    public Triangle(double accessMethodForSquare) {
         super("Triangle", accessMethodForSquare); //κλήση του constructor της υπερκλάσης με τον τύπο "Triangle" και το εμβαδόν του περιβάλλοντος τετραγώνου
     }
